The inliner now looks harder for method impleme...
The inliner now looks harder for method implementations. In addition
to the receiver, it will find those in directly mixed in traits and
in superclasses. It still won't find those in traits mixed only into
superclasses, as that didn't come quite so easily. Closes #3234 but I'll
be opening up another ticket. Review by dragos.