Removes a bunch of private functions which are ...
Removes a bunch of private functions which are never called. While based
on the nature of "private" one can generally feel pretty good that such
a thing is safe, there is always a chance the author had some future use
in mind. On that note I draw your attention in particular to:

  (martin) Typers#stabilizedType: it "sounds" important, but most of it
  has been commented out since 2007 and the little stub part is a never
  called private.

  (iulian) SpecializeTypes#makeTypeArguments: similarly sounds like a
  cornerstone of a transformation until one notices it isn't used.

Unused methods are "attractive nuisances" for anyone (like myself) who
has to figure out how the compiler works by studying the compiler, for
reasons which are no doubt obvious. No review except as noted.