Deprecated the @serializable annotation, introd...
Deprecated the @serializable annotation, introduce a new trait
"scala.Serializable" which has to be extended instead (cross-platform).

Known issues:
 - Companion objects of serializable classes (including case classes) are automatically made serializable. However, they don't extend "Serializable" statically because of the known difficulty (should be done before typing, but hard).
 - Writing "case class C() extends Serializable" gives "error: trait Serializable is inherited twice"
 - Functions are serializable, but don't extend Serializable dynamically (could be fixed by making FunctionN Serializable - shouldn't we?)

Note that @SerialVersionUID continues to be an annotation; it generates
a static field, which is not possible otherwise in scala.

Review by dragos, extempore.

Question to dragos: in JavaPlatform.isMaybeBoxed, why is there a test
for "JavaSerializableClass"? Is that correct?