Making some recent changes less breaky so sbt c...
Making some recent changes less breaky so sbt can be used with trunk
again. Mainly, renamed scala.reflect.internal's AbstractFile to
AbstractFileType. I can't see a way around renaming it and it's becoming
quite a problem not to have sbt. I inspected scala-ide and I don't think
this is going to break it, but in general I don't know how plausible it
is that we can avoid periodic breakage given that we have at least three
source bases (scala, sbt, scala-ide) each with multiple branches with
interdependencies which are not being continuously integrated. Review by
odersky.