Fixed a bug in the optimizer which was preventi...
Fixed a bug in the optimizer which was preventing private methods from
being inlined. Also relaxes a condition related to the "liftedTry"
problem: the inliner has to exclude certain methods from consideration
if there is a value on the stack and the method being inlined has
exception handlers. The new condition is as before, except that it does
not exclude methods of the "try/finally" variety (i.e. finalizers, but
no other exception handlers.) This is necessary to optimize this common
pattern:

  @inline private def foo(body: => Unit) {
    val saved = something
    try body
    finally something = saved
  }

The closure for "body" can be fully eliminated, but only if the contents
of foo can be inlined into the caller.

Closes #4764, review by rompf.