Since trunk seems to be undergoing all kinds of...
Since trunk seems to be undergoing all kinds of changes of the sort
which touch lots of files, the time is ripe for various search/replace
improvements which have stacked up. This one is the replacement of all
"if (settings.debug.value) log" with "debuglog" since as noted in a
recent commit, the inliner will take care of that for you. No review.