Intercept assert and require calls.
And abort calls, and unhandled exceptions, all so I can supplement the
error message with a little of the vast quantity of useful information
which we possess but do not reveal. "Details are sketchy," says the
officer tasked with investigating the crash, but no longer. Also took
the opportunity to eliminate a bunch of one-argument assertions and
requirements if I thought there was any chance I'd someday be facing
them on the wrong end of an incident.

Have you ever dreamed that instead of this experience:

  % scalac -optimise <long list of files>
  error: java.lang.AssertionError: assertion failed: Record Record(anonymous class JavaToScala$$anonfun$makeScalaPackage$1,Map()) does not contain a field value owner$1

Things could proceed more like this:

  % scalac -optimise <long list of files>
  error:
       while compiling:  src/compiler/scala/reflect/runtime/JavaToScala.scala
         current phase:  closelim
       library version:  version 2.10.0.rdev-4267-2012-01-25-gc94d342
      compiler version:  version 2.10.0.rdev-4270-2012-01-26-gd540ddf
    reconstructed args:  -Ydead-code -optimise -Yinline -Yclosure-elim -Yinline-handlers -d /tmp
    error: java.lang.AssertionError: [etc]

You are not dreaming! IT'S ALL HAPPENING