Fix for unnecessary InnerClasses for top-level objects.
Mirror classes (with static forwarders) for top-level
objects would get unnecessary InnerClasses entries
in some cases. See the following comment I introduced
to the place I fixed:

We need to make sure that we do not confuse inner classes
of the class  we mirror with inner classes of the class
we are mirroring. These two sets can be different as
seen in this case:

```scala
 class A {
  class B
  def b: B = new B
 }
 object C extends A
```

Here mirror class of C has a static forwarder for
(inherited) method `b` therefore it refers to class `B`
and needs InnerClasses entry. However, the real class
for `C` (named `C$`) is empty and does not refer to `B`
thus does not need InnerClasses entry it.

NOTE: This logic has been refactored in GenASM and
everything is implemented in a much cleaner way by
having two separate buffers.