Tone down a soft-warning to only show under -Ydebug.
SI-6191 remains open and can lead to incomplete debug
scope information. In 2.10.0, the backend detected when
this happend, and warned the user. But there is little
the user can do about the warning. We have a few test
cases for the problem now, so we should not pollute
the compile output.