SI-8233 Fix regression in backend with boxed nulls
Regressed in SI-7015 / 1b6661b8.

We do need to "unbox" the null (ie, drop a stack from and load
a null) in general. The only time we can avoid this is if the
tree we are adapting is a `Constant(Literal(null))`.

I've added a test for both backends. Only GenICode exhibited
the problem.