SI-8582 emit InnerClasses attribute in GenBCode
I removed the `-bcode` test since we have a build that passes
`-Ybackend:GenBCode` to all tests.

Short intro do the [`InnerClass` attribute][1]:

  - A class needs one `InnerClass` attribute for each of its nested
    classes

  - A class needs the `InnerClass` attribute for all (nested) classes
    that are mentioned in its constant pool

The attribute for a nested class `A$B$C` consists of the long name of
the outer class `A$B`, the short name of the inner class `C`, and
an access flag set describig the visibility.

The attribute seems to be used for reflection.

[1]: http://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.7.6