Handle Dotty Idents
In Scalac any function call has Select as it function. In dotty it could be an Ident.
Eg in Scalac you get Select(Preded$, println);
while in dotty you get Ident(println)
In order to get actual receiver need to either look at the type(which is TermRef) or at owner chain.
First one is more systematic but is isn't exposed in the abstraction API, so using the second one.