Remove PreName
PreName was abusing the implementation of implicit class by using the
underlying implicit conversion. For instance, several method form
Definitions where expecting an argument of type PreName and called with
a String.

This commit completely removes the PreName abstraction by explicitly
calling .toTermName and .toTypeName as required. This change should be
beneficial for both readability (more precise types) and performance.
Indeed by "pushing out" all the calls to .toTermName and .toTypeName to
the call site, these methods will never use virtual dispatch.

To keep the concisness of string literals a couple of methods where
duplicated to take String arguments (and forward to the
TypeName/TermName version).